// MyCompany/Angular/src/app/models/inPurchaseDetail.ts
export class inPurchaseDetail {
    IDDetail: number;
    IDPurchase: number;
    IDProduct: number;
    ProductName: string;
    Volume: number;
    Price: number;
    Amount: number;
    IDFaktur: string;
    IDPartner: number;
    CurrencyPrice: string;
    CurrencyAmount: string;
}