// MyCompany/Angular/src/app/models/tblProduct.ts
export class tblProduct {
    IDProduct: number;
    ProductName: string;
    Description: string;
    Weight: string;
    Size: string;
    IsActive: boolean;
    //IDSupplier: number;
    //SupplierName: string;
    IDCampaign: number;
    //CurrencyBuy: string;
    //CurrencySell: string;
    SearchName: string;
}